Renaming DistributedMemoryRunner.py
